Revert "8222 smooth tool (#8303)"
This reverts commit d9f8fdd0b5ad5bc41ba1663d8e3bc6fcec485a7b.

+++ src/projectscene/view/clipsview/au3/sampledata.h
@@ -2,12 +2,12 @@
 
 namespace au::projectscene {
 struct SampleData {
-    std::vector<double> y {};
-    std::vector<double> x {};
+    std::vector<int> y {};
+    std::vector<int> x {};
 
     SampleData() = default;
 
-    SampleData(std::vector<double> pY, std::vector<double> pX)
+    SampleData(std::vector<int> pY, std::vector<int> pX)
         : y(std::move(pY)), x(std::move(pX)) {}
 
     size_t size() const { return x.size(); }
